Create recipes.sh
I am not posting this to give anyone ideas. I am posting this because I am being misrepresented, I am being denied a fair chance to be heard, and I am assumed to be a criminal and kept under close watch because of things I am not even doing. And I am tired of hearing other people list off the things they are annoyed that they have to deal with and realizing that those things would have been very much appreciated and needed if they had been sent to me and if they had been see for what they were supposed to be and not what they turned into. And as long as I am still dealing with the same cars that follow me around treating me like an enemy until they get right up to me, which is what used to happen, and as long as I have to listen to my neighbors take showers at very odd times of the night that coincide perfectly with when I would expect someone to have mercy and send relief, and as long as I hear people walk by my window and very loudly remind me of how disgusting I am after answering my call for help at someone else's door, and as long as people keep telling me that I am ok when i am not thinking that eventually I will believe it and get to the point that I am so desperate for it all to stop that I will be able to just will everyone's crimes away into oblivion and we would all be happy, then I will keep being a pain in everyone's ass about everything. Because I was the one who had Thunderstrike. I was the one who smelled violets in the ER bay in Marina del Rey when the nurse was confused about me being a real patient, and I was the one who couldn't get help that entire time I was stranded there because people thought I was the devil himself. Literally, I now am coming to realize. I was also the one in Davis when that girl was so excited to see me and I still have no idea why. And that same day I was the one who the shopkeeper accused of being a thief. I had just walked miles in flip flops and was tired and trying to call for help. I had no idea why everyone was so rude, but that is when the mixups started. I have no idea why because I was alone and I never made it anywhere. I ended up in that rest stop, #46, watching scissortails dance around on that scenic overlook. Now it has become even more difficult for me to be seen as who I am, and then under that who I am again. Several times over I was wrapped in disguises it feels like, and it seems like that's what is going on with this code too. Only the code is malicious. I am not. Or maybe it is not either and is just meant to protect everyone else because they think I am evil. But why would anyone consider me evil when my gifts came from Jehovah? If they didn't then I guarantee you there was some kind of malicious and underhanded swap that happened, but you would think that at some point someone somewhere along the way would have told me what it was that made them hate me so much.