Add files via upload
Again, these are not my files. They are from Ubuntu, supposedly, 21.04 that I got in a Linux Magazine on a pressed DVD. I'm sure the DVD is legit, but I'm not sure that the boot is legit since the initrd files I see usually are nothing but microcode inside unless that is some kind of way that they are encoded. Not only that, though, the boot process restarts several times during each boot which is probably why any instructions to avoid certain things passed to the kernel are rendered useless by the time it restarts the third or fourth time. The command line seems to be stored in a variable, though, and placed back in at the end so when you check /proc/cmdline it seems like it is exactly what you instructed it to do. Also sounds familiar to my real life and probably why I get blamed for not functioning correctly.