Update README.md
Added section about /var/log/partman and the 20some_device file. I am trying to make sense of this because it seems waaaayyyyy to complicated for a simple installation. And it takes a long time which makes it so much harder to recover from when other things are quick to reinstall. And when I have to use this one to make sure that everyone knows I am not anti-Africa, that is a little bit frustrating.