Create partman
I think this is from one single installation attempt. It froze when it got to the part where it asked me if I wanted to continue without formatting any of the partitions I had assigned (because I am trying to reinstall it over the previous installation that I broke), but then threw me an error that it couldn't unmount the disk which caused a resize to fail. Then it froze and I force quit it.